Geographic centroid can handle MultiPolygon
